Add Dockerfile for the frontend
The development build works, however there are a couple other things I'd
like to do:
- Run in production mode
  - Use `yarn build`s output and serve it with `serve`
- Run with dumb-init as PID 1

Also need to run the backend in a container and then ensure networking
between the frontend and backend work. I plan on using docker-compose
for this, although not sure if I'll need to mess with docker networking
related configurations? We'll find out